% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu.R
\name{wu.fit}
\alias{wu.fit}
\title{Fit spike and slab Cox PHMs}
\usage{
wu.fit(Y, delta, X)
}
\arguments{
\item{Y}{survival times}

\item{delta}{censoring indicator, 0: uncensored, 1: censored}

\item{X}{design matrix}
}
\value{
list with coefficient values and number of non-zero coefficients
}
\description{
Fit spike and slab Cox PHMs
}
